+++ app/Main.hs
@@ -1,8 +1,12 @@
 module Main where
 
+import App.Arguments
+import Options.Applicative
 import qualified StaticLS.Server as StaticLS
+import StaticLS.StaticEnv.Options
 
 main :: IO ()
 main = do
-    _ <- StaticLS.runServer
+    staticEnvOpts <- execArgParser
+    _ <- StaticLS.runServer staticEnvOpts
     pure ()